[Decomp][Creature] Cohesion split: extract vendor/trainer data to CreatureVendorTrainer.cpp
