+++ src/game/Object/Creature.cpp
@@ -3172,150 +3172,11 @@ uint32 Creature::GetScriptId() const
     return sScriptMgr.GetBoundScriptId(SCRIPTED_UNIT, -int32(GetGUIDLow())) ? sScriptMgr.GetBoundScriptId(SCRIPTED_UNIT, -int32(GetGUIDLow())) : sScriptMgr.GetBoundScriptId(SCRIPTED_UNIT, GetEntry());
 }
 
-/**
- * @brief Gets vendor items directly assigned to this creature entry.
- *
- * @return The vendor item list, or null if none exists.
- */
-VendorItemData const* Creature::GetVendorItems() const
-{
-    return sObjectMgr.GetNpcVendorItemList(GetEntry());
-}
 
-/**
- * @brief Gets vendor items from this creature's vendor template.
- *
- * @return The vendor template item list, or null if none exists.
- */
-VendorItemData const* Creature::GetVendorTemplateItems() const
-{
-    uint32 VendorTemplateId = GetCreatureInfo()->VendorTemplateId;
-    return VendorTemplateId ? sObjectMgr.GetNpcVendorTemplateItemList(VendorTemplateId) : NULL;
-}
 
-/**
- * @brief Gets the current available stock count for a vendor item.
- *
- * @param vItem The vendor item definition.
- * @return The currently available count.
- */
-uint32 Creature::GetVendorItemCurrentCount(VendorItem const* vItem)
-{
-    if (!vItem->maxcount)
-    {
-        return vItem->maxcount;
-    }
 
-    VendorItemCounts::iterator itr = m_vendorItemCounts.begin();
-    for (; itr != m_vendorItemCounts.end(); ++itr)
-    {
-        if (itr->itemId == vItem->item)
-        {
-            break;
-        }
-    }
 
-    if (itr == m_vendorItemCounts.end())
-    {
-        return vItem->maxcount;
-    }
 
-    VendorItemCount* vCount = &*itr;
-
-    time_t ptime = time(NULL);
-
-    if (vCount->lastIncrementTime + vItem->incrtime <= ptime)
-    {
-        ItemPrototype const* pProto = ObjectMgr::GetItemPrototype(vItem->item);
-
-        uint32 diff = uint32((ptime - vCount->lastIncrementTime) / vItem->incrtime);
-        if ((vCount->count + diff * pProto->BuyCount) >= vItem->maxcount)
-        {
-            m_vendorItemCounts.erase(itr);
-            return vItem->maxcount;
-        }
-
-        vCount->count += diff * pProto->BuyCount;
-        vCount->lastIncrementTime = ptime;
-    }
-
-    return vCount->count;
-}
-
-/**
- * @brief Updates and consumes stock count for a limited vendor item.
- *
- * @param vItem The vendor item definition.
- * @param used_count The amount being purchased.
- * @return The remaining count after the update.
- */
-uint32 Creature::UpdateVendorItemCurrentCount(VendorItem const* vItem, uint32 used_count)
-{
-    if (!vItem->maxcount)
-    {
-        return 0;
-    }
-
-    VendorItemCounts::iterator itr = m_vendorItemCounts.begin();
-    for (; itr != m_vendorItemCounts.end(); ++itr)
-    {
-        if (itr->itemId == vItem->item)
-        {
-            break;
-        }
-    }
-
-    if (itr == m_vendorItemCounts.end())
-    {
-        uint32 new_count = vItem->maxcount > used_count ? vItem->maxcount - used_count : 0;
-        m_vendorItemCounts.push_back(VendorItemCount(vItem->item, new_count));
-        return new_count;
-    }
-
-    VendorItemCount* vCount = &*itr;
-
-    time_t ptime = time(NULL);
-
-    if (vCount->lastIncrementTime + vItem->incrtime <= ptime)
-    {
-        ItemPrototype const* pProto = ObjectMgr::GetItemPrototype(vItem->item);
-
-        uint32 diff = uint32((ptime - vCount->lastIncrementTime) / vItem->incrtime);
-        if ((vCount->count + diff * pProto->BuyCount) < vItem->maxcount)
-        {
-            vCount->count += diff * pProto->BuyCount;
-        }
-        else
-        {
-            vCount->count = vItem->maxcount;
-        }
-    }
-
-    vCount->count = vCount->count > used_count ? vCount->count - used_count : 0;
-    vCount->lastIncrementTime = ptime;
-    return vCount->count;
-}
-
-/**
- * @brief Gets trainer spells from the trainer template bound to this creature.
- *
- * @return The trainer template spell list, or null if none exists.
- */
-TrainerSpellData const* Creature::GetTrainerTemplateSpells() const
-{
-    uint32 TrainerTemplateId = GetCreatureInfo()->TrainerTemplateId;
-    return TrainerTemplateId ? sObjectMgr.GetNpcTrainerTemplateSpells(TrainerTemplateId) : NULL;
-}
-
-/**
- * @brief Gets trainer spells directly assigned to this creature entry.
- *
- * @return The trainer spell list, or null if none exists.
- */
-TrainerSpellData const* Creature::GetTrainerSpells() const
-{
-    return sObjectMgr.GetNpcTrainerSpells(GetEntry());
-}
 
 // overwrite WorldObject function for proper name localization
 /**
